# StarkSavingsBank-Backend/src/main/resources/static/openapi.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Stark Savings Bank API Documentation"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
...
    "paths": {
...
        "/reset-password": {
            "post": {
                "tags": [
                    "Reset-Password API"
                ],
                "description": "Saves the new password",
                "operationId": "resetPassword",
                "requestBody": {
                    "description": "Password reset request DTO",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ResetRequestDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/password-forgot": {
            "post": {
                "tags": [
                    "Reset-Password API"
                ],
                "description": "Send reset password email",
                "operationId": "sendResetPasswordMail",
                "requestBody": {
                    "description": "Email Reset Request Object",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EmailResetRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/contact": {
            "post": {
                "tags": [
                    "Contacts API"
                ],
                "description": "Create a new contact message",
                "operationId": "saveMessage",
                "requestBody": {
                    "description": "contact message object",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Contact"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Contact"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Customers API"
                ],
                "description": "Authenticates the user",
                "operationId": "loginApi",
                "requestBody": {
                    "description": "Login Request Data Transfer Object",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequestDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/LoginResponseDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user": {
            "get": {
                "tags": [
                    "Customers API"
                ],
                "description": "Check if user already exists in the database",
                "operationId": "getUserDetails",
                "requestBody": {
                    "description": "Authentication Object",
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "string"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Customer"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "EmailResetRequest": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string"
                    }
                }
            },
...
            "ResetRequestDTO": {
                "type": "object",
                "properties": {
                    "password": {
                        "type": "string"
                    },
                    "confirmPassword": {
                        "type": "string"
                    }
                }
            },
            "Customer": {
                "type": "object",
                "properties": {
                    "name": {
                        "maxLength": 100,
                        "minLength": 5,
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "mobileNumber": {
                        "type": "string"
                    },
                    "houseAffiliation": {
                        "type": "string"
                    },
                    "role": {
                        "type": "string",
                        "enum": [
                            "ADMIN",
                            "CUSTOMER",
                            "EMPLOYEE"
                        ]
                    },
                    "createDate": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "password": {
                        "type": "string",
                        "writeOnly": true
                    },
                    "confirmPassword": {
                        "type": "string",
                        "writeOnly": true
                    }
                }
            },
            "Contact": {
                "type": "object",
                "properties": {
                    "contactId": {
                        "type": "string"
                    },
                    "contactName": {
                        "type": "string"
                    },
                    "contactEmail": {
                        "type": "string"
                    },
                    "house": {
                        "type": "string"
                    },
                    "subject": {
                        "maxLength": 100,
                        "minLength": 5,
                        "type": "string"
                    },
                    "message": {
                        "maxLength": 200,
                        "minLength": 5,
                        "type": "string"
                    }
                }
            },
            "LoginRequestDTO": {
                "type": "object",
                "properties": {
                    "username": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    }
                }
            },
            "LoginResponseDTO": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "string"
                    },
                    "jwt": {
                        "type": "string"
                    }
                }
            },
...
        }
    }
}